reorganize encryption error types
